Finish splitting up main parser
